requirements.txt
Django==2.2
Faker==1.0.5
Pillow==6.0.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-jquery==3.1.0